requirements.txt
APScheduler==3.11.2
asgiref==3.10.0
attrs==25.4.0
Django==5.2.7
django-apscheduler==0.7.0
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.0.0
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
